requirements.txt
django-extensions
ipython
notebook
Pillow
pilkit
django-imagekit
channels
asgi_redis
